+++ src/komorebi/blog.py
@@ -15,7 +15,8 @@ from flask import (
 from flask_httpauth import HTTPBasicAuth
 from passlib.apache import HtpasswdFile
 
-from . import db, embeds, formatting, forms, time
+from . import db, embeds, formatting, forms
+from .adjunct import time
 from .feed import generate_feed
 
 blog = Blueprint("blog", __name__)
